close panel when click outside
