+++ src/pages/Dashboard.tsx
@@ -3,7 +3,6 @@ import {
   Backdrop,
   Box,
   Button,
-  Card,
   CircularProgress,
   Dialog,
   DialogActions,
@@ -17,8 +16,6 @@ import {
 import { FC, useEffect, useState } from 'react';
 import { SummaryTable } from '../components/SummaryTable/SummaryTable';
 import { useLoaderData } from 'react-router-dom';
-import { TableSliderFilter } from '../components/Filters/TableSliderFilter';
-import { TableOptionFilter } from '../components/Filters/TableOptionFilter';
 import axios from 'axios';
 import { Filter, brainDataFilters } from '../types/Filter';
 import { ExpandableChip } from '../components/ExpandableChip';
@@ -36,9 +33,6 @@ export const DashboardPage: FC = () => {
   const [filterRequest, setFilterRequest] = useState<FilterRequest>({ categories: {} });
   const [loading, setLoading] = useState(false);
   const [displayClearFilters, setDisplayClearFilters] = useState(false);
-
-  // filters dropdown states
-  const [filterDropdowns, setFilterDropdowns] = useState<{ [key: string]: boolean }>({});
 
   const [selectedCategories, setSelectedCategories] = useState<string[]>([]);
   const [inputValue, setInputValue] = useState<string | undefined>('');
@@ -78,19 +72,6 @@ export const DashboardPage: FC = () => {
         return newState;
       });
     }
-  };
-
-  const handleFilterDropdown = (name: string) => {
-    // set all dropdowns to false except the one clicked
-    setFilterDropdowns((prevState) => {
-      const newState = Object.keys(prevState).reduce((acc, key) => {
-        if (key !== name) acc[key] = false;
-        return acc;
-      }, {} as { [key: string]: boolean });
-      newState[name] = !prevState[name];
-
-      return newState;
-    });
   };
 
   const handleAddFilter = (_event: any, newValue: string[]) => {
@@ -153,16 +134,6 @@ export const DashboardPage: FC = () => {
     if (filters.length > 1) setDisplayClearFilters(true);
     else setDisplayClearFilters(false);
 
-    // set all categories to false
-    setFilterDropdowns((prevState) => {
-      const newState = Object.keys(prevState).reduce((acc, key) => {
-        acc[key] = false;
-        return acc;
-      }, {} as { [key: string]: boolean });
-
-      return newState;
-    });
-
     // add new filter to filter request
     setFilterRequest((prevState) => {
       const newState: FilterRequest = { ...prevState, categories: prevState.categories ?? {} };
@@ -217,62 +188,16 @@ export const DashboardPage: FC = () => {
             </Button>
           )}
           <Box display="flex" alignItems="center" flexWrap="wrap" gap={1}>
-            {filters.map((filter) => {
-              // add filter value to chip extended label
-              let expandText = filter.variableName;
-
-              // get filter values
-              const filterValues = filter.npCategory ? (filterRequest.categories[filter.name] as number[]) : (filterRequest[filter.name] as number[]);
-
-              if (filterValues) {
-                if (filter.type === 'slider') {
-                  expandText += `: ${filterValues[0]} - ${filterValues[1]}`;
-                } else if (filter.type === 'option') {
-                  const filterLabels = [];
-                  for (const label of Object.keys(filter.options)) {
-                    if (filterValues.includes(filter.options[label])) filterLabels.push(label);
-                  }
-                  expandText += `: ${filterLabels.join(', ')}`;
-                }
-              }
-
-              return (
-                <div key={filter.name}>
-                  <ExpandableChip
-                    label={`${filter.variableName}`}
-                    expendedLabel={expandText}
-                    onClick={() => handleFilterDropdown(filter.variableName)}
-                    onDelete={() => handleRemoveFilter(filter.name, filter.variableName, filter.npCategory)}
-                  />
-                  {filterDropdowns[filter.variableName] && (
-                    <Box component={Card} zIndex={1} position="absolute" padding={2} minWidth={300}>
-                      {filter.type === 'slider' ? (
-                        <TableSliderFilter
-                          filterName={filter.name}
-                          variableName={filter.variableName}
-                          npCatagory={filter.npCategory}
-                          maxValue={filter.max}
-                          minValue={filter.min}
-                          step={filter.step}
-                          value={filterValues}
-                          applyFilter={changeFilter}
-                        />
-                      ) : (
-                        <TableOptionFilter
-                          filterName={filter.name}
-                          variableName={filter.variableName}
-                          npCatagory={filter.npCategory}
-                          optionType={filter.optionType}
-                          options={filter.options}
-                          values={filterValues}
-                          applyFilter={changeFilter}
-                        />
-                      )}
-                    </Box>
-                  )}
-                </div>
-              );
-            })}
+            {filters.map((filter) => (
+              <ExpandableChip
+                key={filter.name}
+                filter={filter}
+                filterRequest={filterRequest}
+                label={`${filter.variableName}`}
+                applyFilter={changeFilter}
+                onDelete={() => handleRemoveFilter(filter.name, filter.variableName, filter.npCategory)}
+              />
+            ))}
           </Box>
         </Box>
         <Divider sx={{ m: 2 }} />
